SatMapModifier/requirements.txt
# Requirements for SatMapModifier
# Image processing library
Pillow>=10.0.0
opencv-python>=4.7.0
numpy>=1.25.0